finish

+++ app/controllers/user_courses_controller.rb
@@ -1,7 +1,8 @@
 class UserCoursesController < ApplicationController
   before_action :logged_in_user
-  before_action :load_course, only: %i(index create del_user_courses)
+  before_action :load_course, only: %i(index create del_user_courses destroy)
   before_action :load_user_course, only: :destroy
+  before_action :verify_can_modify, only: %i(create destroy)
 
   def index
     @users_system = User.belongs_to_suppervisor params[:option]
@@ -10,38 +11,46 @@ class UserCoursesController < ApplicationController
   end
 
   def create
-    if params[:ids].present?
-      @user_courses = Array.new
-      @ids_deny = Array.new
-      params[:ids].each do |user|
-        load_user user
-        user_course = @course.user_courses.create user_id: @user.id
-        if user_course
-          if @user.trainee?
-            add_user_subject @course, @user
+    if @permited
+      if params[:ids].present?
+        @user_courses = Array.new
+        @ids_deny = Array.new
+        params[:ids].each do |user|
+          load_user user
+          user_course = @course.user_courses.create user_id: @user.id
+          if user_course
+            if @user.trainee?
+              add_user_subject @course, @user
+            end
+            @user_courses << user_course
+          else
+            @ids_deny << user
           end
-          @user_courses << user_course
-        else
-          @ids_deny << user
         end
+        @mes_success = "Insert user into course successfully"
+      else
+        @mes_success = "Nothing changes"
       end
-      @mes_success = "Insert user into course successfully"
     else
-      @mes_success = "Nothing changes"
+      @mes_fail = "This course must be init or in progress"
     end
   end
 
   def destroy
-    if @user_course.destroy
-      if @user_course.user.trainee?
-        course_subjects = @user_course.course.course_subjects
-        course_subjects.each do |cs|
-          @user_course.user.user_subjects.find_by(course_subject_id: cs.id).destroy
+    if @permited
+      if @user_course.destroy
+        if @user_course.user.trainee?
+          course_subjects = @user_course.course.course_subjects
+          course_subjects.each do |cs|
+            @user_course.user.user_subjects.find_by(course_subject_id: cs.id).destroy
+          end
         end
+        @mes_success = "Remove user from course successfully"
+      else
+        @mes_danger = "Remove user from course unsuccess"
       end
-      @mes_success = "Remove user from course successfully"
     else
-      @mes_danger = "Remove user from course unsuccess"
+      @mes_fail = "This course must be init or in progress"
     end
   end
 
@@ -55,28 +64,32 @@ class UserCoursesController < ApplicationController
   end
 
   def del_user_courses
-    if params[:ids].present?
-      @ids_deny = Array.new
-      @users = Array.new
-      params[:ids].each do |line|
-        user_cache = UserCourse.find_by(id: line).user
-        @uc = UserCourse.find_by(id: line)
-        if UserCourse.find_by(id: line).destroy
-          @users << user_cache
-          course = @uc.course
-          if user_cache.trainee?
-            course_subjects = course.course_subjects
-            course_subjects.each do |cs|
-              user_cache.user_subjects.find_by(course_subject_id: cs.id).destroy
+    if @permited
+      if params[:ids].present?
+        @ids_deny = Array.new
+        @users = Array.new
+        params[:ids].each do |line|
+          user_cache = UserCourse.find_by(id: line).user
+          @uc = UserCourse.find_by(id: line)
+          if UserCourse.find_by(id: line).destroy
+            @users << user_cache
+            course = @uc.course
+            if user_cache.trainee?
+              course_subjects = course.course_subjects
+              course_subjects.each do |cs|
+                user_cache.user_subjects.find_by(course_subject_id: cs.id).destroy
+              end
             end
+          else
+            @ids_deny << line
           end
-        else
-          @ids_deny << line
         end
+        @mes_success = "Remove users from course successfully"
+      else
+        @mes_success = "Nothing changes"
       end
-      @mes_success = "Remove users from course successfully"
     else
-      @mes_success = "Nothing changes"
+      @mes_fail = "This course must be init or in progress"
     end
   end
 
@@ -111,5 +124,13 @@ class UserCoursesController < ApplicationController
     return if @user_course
     flash[:danger] = "There was an error"
     redirect_to root_path
+  end
+
+  def verify_can_modify
+    if @course.finish? || @course.block?
+      @permited = false
+    else
+      @permited = true
+    end
   end
 end
